Tom Rush Main Menu Edits
I created a main menu level with three buttons, Start, Options, and Quit. Start transfers the player to the TopDownMap (we can change that in the future depending on the other level builds). Options does not go anywhere (if we want to implement an options page we can, but it can be done later). Quit exits the game.